Fix warning: mysqli_free_result(): Couldn't fetch mysqli_result
The problem was a double-free on the result variable. Solved by making
the MySQLiResultSet's handle variable a reference to the result instead
of holding a copy.

To reproduce, run an unbuffered query and a buffered one without reading
all results from the previous was necessary. The first result *also* needed
to be garbage-collected *before* the subsequent query ran to cause the
double-free